update from cg023

+++ Silverware/src/drv_softi2c.h
@@ -2,11 +2,11 @@
 
 
 void softi2c_init(void);
-void softi2c_readdata(uint8_t device_address ,uint8_t register_address , int *data, int size );
-void softi2c_writedata(uint8_t device_address ,uint8_t register_address , int *data, int size );
+void softi2c_readdata(int device_address ,int register_address , int *data, int size );
+void softi2c_writedata(int device_address ,int register_address , int *data, int size );
 
-uint8_t softi2c_read(uint8_t device_address , uint8_t register_address);
-uint8_t softi2c_write( uint8_t device_address , uint8_t address,uint8_t value);
+int softi2c_read(int device_address , int register_address);
+int softi2c_write( int device_address , int address,int value);
